Added page objects / tests for Amazon.in Base Class - Generic functions updated as per the new locator strategy conftest.py - commented line#86 Updated ReadMe file Added EmailUtility class file Updated conftest.py to add location of javascript folder. Updated conf.json to add details for email Added new folder js_scripts and moved the drag_and_drop_helper.js file to that path.
Added the locator values in FunctionalPageObjects.py class
Added seperate package for Amazon Pages and Amazon Locators(page wise)
