+++ page_objects/amazon_objects/amazon_homepage_objects.py
@@ -1,0 +1,4 @@
+# Amazon
+homepage_searchbox = "xpath://input[@id='twotabsearchtextbox']"
+homepage_findbutton = "xpath://input[@id='nav-search-submit-button']"
+
