+++ page_objects/amazon_objects/amazon_productspage_objects.py
@@ -1,0 +1,8 @@
+
+add_to_cart = "id:add-to-cart-button"
+cart_button = "xpath://*[@id='attach-sidesheet-view-cart-button']/span/input"
+quantity = "xpath://select[@name='quantity']"
+second_object_recent_history = "xpath://*[@id='a-autoid-12']/span/input"
+goto_cart = "xpath://*[@id='sw-gtc']/span/a"
+first_product_in_cart = "xpath://div[@data-name='Active Items']/div[3]/div[4]/div/div[2]/div[1]/span[2]/span/input"
+proceed_to_checkout = "xpath://*[@value='Proceed to checkout']"
