+++ page_objects/amazon_objects/amazon_searchresultpage_objects.py
@@ -1,0 +1,1 @@
+product_name = "xpath://*[text()='Apple iPhone 14 (128 GB) - Purple']"
